refactor(view): modularize side-by-side views
